perf: exactitud de notas, boletin sin N+1 e identidad memorizada (Fase 8)
Cuatro asuntos acumulados de fases anteriores.

V14 pasa las notas de DOUBLE a DECIMAL(3,2) y el codigo a BigDecimal.
DOUBLE es binario y no representa exactamente 2.9 ni 3.05; RB-12 decide
la aprobacion comparando contra 3.0, de modo que una nota guardada como
2.9999999999999996 cae del lado equivocado de una comparacion que decide
si alguien pierde el ano. El dominio normaliza a dos decimales y compara
con compareTo, no con equals, porque 3.0 y 3.00 son el mismo numero. En
el promedio ponderado los aportes se acumulan sin redondear y solo se
redondea el total. El porcentaje de asistencia se queda en double a
proposito: es un porcentaje calculado, no un valor de la escala.

El boletin hacia N+1 consultas desde la propia Fase 6: notas,
ponderaciones y asistencia por cada materia del plan. Ahora son tres, y
el calculo se separo de la carga de datos para poder resolver todas las
materias con lo ya traido.

ContextoUsuario memoriza la identidad por peticion: una sola llamada a
puedeVerEstudiante leia la cuenta cuatro veces. Es un ThreadLocal y no un
bean @RequestScope porque el contexto tambien se usa desde listeners,
donde el ambito de peticion no existe. Lo importante no es el ahorro sino
la limpieza: el servidor reutiliza hilos, asi que la memoria va marcada
con el correo y LimpiezaIdentidadFilter la vacia en un finally.

La recuperacion de contrasena limita las solicitudes por cuenta y
ventana, contando las filas que ya existen en token_recuperacion en vez
de llevar estado aparte. Corta en silencio: responder "demasiadas
peticiones" confirmaria que ese correo tiene cuenta.

+++ backend/src/main/java/com/educktrack/notas/infrastructure/rest/NotaDtos.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -38,31 +39,31 @@ public final class NotaDtos {
             @NotNull Long cursoId,
             @NotNull Long periodoAcademicoId,
             @NotNull(message = "El tipo de evaluacion es obligatorio") TipoEvaluacion tipo,
-            @NotNull(message = "El valor es obligatorio") Double valor,
+            @NotNull(message = "El valor es obligatorio") BigDecimal valor,
             String descripcion) {
     }
 
-    public record EditarCalificacionRequest(@NotNull Double valor) {
+    public record EditarCalificacionRequest(@NotNull BigDecimal valor) {
     }
 
     public record CalificacionDto(
             Long id, Long estudianteId, Long materiaId, Long cursoId, Long periodoAcademicoId,
-            TipoEvaluacion tipo, double valor, String descripcion) {
+            TipoEvaluacion tipo, BigDecimal valor, String descripcion) {
     }
 
     // ---------- Promedio (RF-33) ----------
-    public record DetalleTipoDto(TipoEvaluacion tipo, Double promedioTipo, int porcentaje, double aporte) {
+    public record DetalleTipoDto(TipoEvaluacion tipo, BigDecimal promedioTipo, int porcentaje, BigDecimal aporte) {
     }
 
     public record PromedioDto(
             Long estudianteId, Long materiaId, Long periodoAcademicoId,
-            double promedio, boolean aprobada,
+            BigDecimal promedio, boolean aprobada,
             List<TipoEvaluacion> pendientes, List<DetalleTipoDto> detalle) {
     }
 
     // ---------- Novedad (RF-36, RB-15) ----------
     public record NovedadRequest(
-            @NotNull(message = "El nuevo valor es obligatorio") Double nuevoValor,
+            @NotNull(message = "El nuevo valor es obligatorio") BigDecimal nuevoValor,
             @NotBlank(message = "El motivo es obligatorio") String motivo) {
     }
 
@@ -80,12 +81,12 @@ public final class NotaDtos {
      * minimo de asistencia sigue siendo aprobatoria, y quien lee el boletin
      * necesita ver las dos cosas para decidir.</p>
      */
-    public record BoletinMateriaDto(Long materiaId, double promedio, boolean aprobada,
+    public record BoletinMateriaDto(Long materiaId, BigDecimal promedio, boolean aprobada,
                                     boolean sinCalificar, boolean pierdeDerechoAEvaluacion) {
     }
 
     public record BoletinDto(
             Long estudianteId, Long cursoId, Long periodoAcademicoId,
-            List<BoletinMateriaDto> materias, double promedioGeneral, boolean aprobado) {
+            List<BoletinMateriaDto> materias, BigDecimal promedioGeneral, boolean aprobado) {
     }
 }
